+++ R/project.R
@@ -8,17 +8,42 @@
 #' @param WEIGHT The sample weight for the sample file. (So a 2\% file has \code{WEIGHT} = 50.)
 #' @param excl_vars A character vector of column names in \code{sample_file} that should not be inflated. Columns not present in the 2013-14 sample file are not inflated and nor are the columns \code{Ind}, \code{Gender}, \code{age_range}, \code{Occ_code}, \code{Partner_status}, \code{Region}, \code{Lodgment_method}, and \code{PHI_Ind}.
 #' @param forecast.dots A list containing parameters to be passed to \code{generic_inflator}.
-#' @param wage.series See \code{\link{wage_inflator}}. Note that the \code{Sw_amt} will uprated by \code{\link{differentially_uprate_wage}}.
+#' @param wage.series See \code{\link{wage_inflator}}. Note that the \code{Sw_amt}
+#'  will uprated by \code{\link{differentially_uprate_wage}} (if requested).
 #' @param lf.series See \code{\link{lf_inflator_fy}}.
 #' @param use_age_pop_forecast Should the inflation of the number of taxpayers be 
 #' moderated by the number of resident persons born in a certain year? If \code{TRUE},
 #' younger ages will grow at a slightly higher rate beyond 2018 than older ages.
-#' @param .recalculate.inflators (logical, default: \code{FALSE}. Should \code{generic_inflator()} or \code{CG_inflator} be called to project the other variables? Adds time.
-#' @param .copyDT (logical, default: \code{TRUE}) Should a \code{copy()} of \code{sample_file} be made? If set to \code{FALSE}, will update \code{sample_file}. 
-#' @param check_fy_sample_file (logical, default: \code{TRUE}) Should \code{fy.year.of.sample.file} be checked against \code{sample_file}?
-#' By default, \code{TRUE}, an error is raised if the base is not 2012-13, 2013-14, 2014-15, 2015-16, or 2016-17, and a warning is raised if the 
+#' @param .recalculate.inflators (logical, default: \code{NA}). 
+#' Should \code{generic_inflator()} or \code{CG_inflator} be called to project the other variables? Adds time.
+#' Default \code{NA} means \code{TRUE} if the pre-calculated inflators are available,
+#' \code{FALSE} otherwise.
+#' 
+#' @param .copyDT (logical, default: \code{TRUE}) Should a \code{copy()} of \code{sample_file} be made?
+#'  If set to \code{FALSE}, will update \code{sample_file} in place, which may
+#'  be necessary when memory is constrained, but is dangerous as it modifies the
+#'  original data and its projection. (So if you run the same code twice you 
+#'  may end up with a projection \code{2h} years ahead, not \code{h} years.)
+#' @param check_fy_sample_file (logical, default: \code{TRUE}) 
+#' Should \code{fy.year.of.sample.file} be checked against \code{sample_file}?
+#' By default, \code{TRUE}, an error is raised if the base is not 2012-13, 2013-14, 2014-15, 2015-16, 2016-17,
+#' or 2017-18,
+#'  and a warning is raised if the 
 #' number of rows in \code{sample_file} is different to the known number of rows in the sample files. 
-#' @param differentially_uprate_Sw (logical, default: \code{TRUE}) Should the salary and wage column (\code{Sw_amt}) be differentially uprating using (\code{\link{differentially_uprate_wage}})?
+#' @param differentially_uprate_Sw (logical, default: \code{NA}) 
+#' Should the salary and wage column (\code{Sw_amt}) be differentially uprated 
+#' using (\code{\link{differentially_uprate_wage}})? Default of \code{NA} means
+#' use differential uprating is used when \code{fy.year.of.sample.file <= "2016-17"}.
+#' It is known that the Treasury stopped using differential uprating by 2019.
+#' 
+#' Selecting \code{TRUE} for \code{fy.year.of.sample.file > "2016-17"} is an 
+#' error as the precalculated values are not available.
+#' 
+#' @param r_super_balance The factor to inflate super balances by (annualized).
+#' Set to \code{1.05} for backwards compatibility. The annual superannuation 
+#' bulletin of June 2019 from APRA reported 7.3\% growth of funds with more than
+#' fund members over the previous 5 years and 7.9\% growth over the 
+#' previous ten years.
 #' 
 #' 
 #' @return A sample file with the same number of rows as \code{sample_file} but 
@@ -46,7 +71,7 @@
 #' as the latter suggests lower-than-recorded tax collections. 
 #' However, more recent data is of course preferable.
 #' @examples 
-#' # install.packages('taxstats', repos = 'https://hughparsonage.github.io/drat')
+#' # install_taxstats()
 #' if (requireNamespace("taxstats", quietly = TRUE) &&
 #'     requireNamespace("data.table", quietly = TRUE)) {
 #'   library(taxstats)
@@ -67,10 +92,11 @@ project <- function(sample_file,
                     wage.series = NULL,
                     lf.series = NULL,
                     use_age_pop_forecast = FALSE,
-                    .recalculate.inflators = FALSE, 
+                    .recalculate.inflators = NA, 
                     .copyDT = TRUE,
                     check_fy_sample_file = TRUE,
-                    differentially_uprate_Sw = TRUE) {
+                    differentially_uprate_Sw = NA,
+                    r_super_balance = 1.05) {
   if (length(h) != 1L) {
     stop("`h` had length-", length(h), ", ", 
          "but must be a length-1 positive integer.")
@@ -190,6 +216,7 @@ project <- function(sample_file,
   # NSE e.g inflators[h == h]
   H <- h
   current.fy <- fy.year.of.sample.file
+  
   to.fy <- yr2fy(fy2yr(current.fy) + h)
   
   if (is.null(wage.series)){
@@ -209,7 +236,20 @@ project <- function(sample_file,
   
   cpi.inflator <- cpi_inflator(1, from_fy = current.fy, to_fy = to.fy)
   
-  if (.recalculate.inflators || current.fy >= "2017-18") {
+  if (!is.logical(.recalculate.inflators)) {
+    stop("`.recalculate.inflators` was type ", typeof(.recalculate.inflators), ", ",
+         "but must be TRUE, FALSE, or NA.")
+  }
+  if (length(.recalculate.inflators) != 1L) {
+    stop("`length(.recalculate.inflators) = ", length(.recalculate.inflators), ", 
+         but must be length-one.")
+  }
+  
+  if (is.na(.recalculate.inflators)) {
+    .recalculate.inflators <- current.fy >= "2017-18"
+  }
+  
+  if (.recalculate.inflators) {
     CG.inflator <- CG_inflator(1, from_fy = current.fy, to_fy = to.fy)
   } else {
     if (current.fy %notin% c("2012-13", "2013-14", 
@@ -235,6 +275,10 @@ project <- function(sample_file,
   col.names <- names(sample_file)
   
   # Differential uprating not available for years outside:
+  if (is.na(differentially_uprate_Sw)) {
+    differentially_uprate_Sw <- fy.year.of.sample.file <= "2016-17"
+  }
+  
   diff.uprate.wagey.cols <- 
     if (differentially_uprate_Sw) {
       "Sw_amt"
@@ -375,6 +419,15 @@ project <- function(sample_file,
     lf.inflator <- {lf.inflator + .subset2(sample_file, "pop_forecast_r")} / 2
   }
   
+  if (!is.numeric(r_super_balance)) {
+    stop("`r_super_balance` was type ", typeof(r_super_balance), " but must be
+         numeric.")
+  }
+  if (length(r_super_balance) != 1L && length(r_super_balance) != nrow(sample_file)) {
+    stop("`length(r_super_balance) = ", length(r_super_balance), "`, but must be 
+         length-one or have the same length as the number of rows of `sample_file = ",
+         nrow(sample_file), "`.")
+  }
 
   for (j in col.names) {
     if (j %chin% Not.Inflated) {
@@ -416,7 +469,7 @@ project <- function(sample_file,
                }
              },
              "super" = {
-               {1.05 ^ h} * v
+               {r_super_balance ^ h} * v
              },
              # nocov start
              stop("Internal error: switch(inflator_switch()) not matched\n\t",